// packages/thrift-client/src/tests/integration/generated-apache/shared/shared.ts
export interface ISharedStructArgs {
    key: number;
    value: string;
}
export class SharedStruct {
    public key: number;
    public value: string;
    constructor(args: ISharedStructArgs) {
        if (args != null && args.key != null) {
            this.key = args.key;
        }
        else {
            throw new thrift.Thrift.TProtocolException(thrift.Thrift.TProtocolExceptionType.UNKNOWN, "Required field key is unset!");
        }
        if (args != null && args.value != null) {
            this.value = args.value;
        }
        else {
            throw new thrift.Thrift.TProtocolException(thrift.Thrift.TProtocolExceptionType.UNKNOWN, "Required field value is unset!");
        }
    }
    public write(output: thrift.TProtocol): void {
        output.writeStructBegin("SharedStruct");
        if (this.key != null) {
            output.writeFieldBegin("key", thrift.Thrift.Type.I32, 1);
            output.writeI32(this.key);
            output.writeFieldEnd();
        }
        if (this.value != null) {
            output.writeFieldBegin("value", thrift.Thrift.Type.STRING, 2);
            output.writeString(this.value);
            output.writeFieldEnd();
        }
        output.writeFieldStop();
        output.writeStructEnd();
        return;
    }
    public static read(input: thrift.TProtocol): SharedStruct {
        input.readStructBegin();
        let _args: any = {};
        while (true) {
            const ret: thrift.TField = input.readFieldBegin();
            const fieldType: thrift.Thrift.Type = ret.ftype;
            const fieldId: number = ret.fid;
            if (fieldType === thrift.Thrift.Type.STOP) {
                break;
            }
            switch (fieldId) {
                case 1:
                    if (fieldType === thrift.Thrift.Type.I32) {
                        const value_1: number = input.readI32();
                        _args.key = value_1;
                    }
                    else {
                        input.skip(fieldType);
                    }
                    break;
                case 2:
                    if (fieldType === thrift.Thrift.Type.STRING) {
                        const value_2: string = input.readString();
                        _args.value = value_2;
                    }
                    else {
                        input.skip(fieldType);
                    }
                    break;
                default: {
                    input.skip(fieldType);
                }
            }
            input.readFieldEnd();
        }
        input.readStructEnd();
        if (_args.key !== undefined && _args.value !== undefined) {
            return new SharedStruct(_args);
        }
        else {
            throw new thrift.Thrift.TProtocolException(thrift.Thrift.TProtocolExceptionType.UNKNOWN, "Unable to read SharedStruct from input");
        }
    }
}
export interface ISharedUnionArgs {
    option1?: string;
    option2?: string;
}
export class SharedUnion {
    public option1?: string;
    public option2?: string;
    constructor(args?: ISharedUnionArgs) {
        let _fieldsSet: number = 0;
        if (args != null) {
            if (args.option1 != null) {
                _fieldsSet++;
                this.option1 = args.option1;
            }
            if (args.option2 != null) {
                _fieldsSet++;
                this.option2 = args.option2;
            }
            if (_fieldsSet > 1) {
                throw new thrift.Thrift.TProtocolException(thrift.Thrift.TProtocolExceptionType.INVALID_DATA, "Cannot read a TUnion with more than one set value!");
            }
            else if (_fieldsSet < 1) {
                throw new thrift.Thrift.TProtocolException(thrift.Thrift.TProtocolExceptionType.INVALID_DATA, "Cannot read a TUnion with no set value!");
            }
        }
    }
    public static fromOption1(option1: string): SharedUnion {
        return new SharedUnion({ option1 });
    }
    public static fromOption2(option2: string): SharedUnion {
        return new SharedUnion({ option2 });
    }
    public write(output: thrift.TProtocol): void {
        output.writeStructBegin("SharedUnion");
        if (this.option1 != null) {
            output.writeFieldBegin("option1", thrift.Thrift.Type.STRING, 1);
            output.writeString(this.option1);
            output.writeFieldEnd();
        }
        if (this.option2 != null) {
            output.writeFieldBegin("option2", thrift.Thrift.Type.STRING, 2);
            output.writeString(this.option2);
            output.writeFieldEnd();
        }
        output.writeFieldStop();
        output.writeStructEnd();
        return;
    }
    public static read(input: thrift.TProtocol): SharedUnion {
        let _fieldsSet: number = 0;
        let _returnValue: SharedUnion | null = null;
        input.readStructBegin();
        while (true) {
            const ret: thrift.TField = input.readFieldBegin();
            const fieldType: thrift.Thrift.Type = ret.ftype;
            const fieldId: number = ret.fid;
            if (fieldType === thrift.Thrift.Type.STOP) {
                break;
            }
            switch (fieldId) {
                case 1:
                    if (fieldType === thrift.Thrift.Type.STRING) {
                        _fieldsSet++;
                        const value_3: string = input.readString();
                        _returnValue = SharedUnion.fromOption1(value_3);
                    }
                    else {
                        input.skip(fieldType);
                    }
                    break;
                case 2:
                    if (fieldType === thrift.Thrift.Type.STRING) {
                        _fieldsSet++;
                        const value_4: string = input.readString();
                        _returnValue = SharedUnion.fromOption2(value_4);
                    }
                    else {
                        input.skip(fieldType);
                    }
                    break;
                default: {
                    input.skip(fieldType);
                }
            }
            input.readFieldEnd();
        }
        input.readStructEnd();
        if (_fieldsSet > 1) {
            throw new thrift.Thrift.TProtocolException(thrift.Thrift.TProtocolExceptionType.INVALID_DATA, "Cannot read a TUnion with more than one set value!");
        }
        else if (_fieldsSet < 1) {
            throw new thrift.Thrift.TProtocolException(thrift.Thrift.TProtocolExceptionType.INVALID_DATA, "Cannot read a TUnion with no set value!");
        }
        if (_returnValue !== null) {
            return _returnValue;
        }
        else {
            throw new thrift.Thrift.TProtocolException(thrift.Thrift.TProtocolExceptionType.UNKNOWN, "Unable to read data for TUnion");
        }
    }
}